Module renaming & small fixes
