+++ dpnp/backend/extensions/statistics/__init__.py
@@ -1,0 +1,1 @@
+from ._statistics_impl import histogram
