+++ setup.py
@@ -66,6 +66,7 @@ setup(
         "dpnp.fft",
         "dpnp.linalg",
         "dpnp.random",
+        "dpnp.backend.extensions.statistics",
     ],
     package_data={
         "dpnp": [
